Everything now runs as expected, successfully parsing UI components into scenes.

+++ src/Building.cpp
@@ -68,7 +68,7 @@ Cuboid& Building::getCuboidRef(int index)
     return (*this->boxes)[index];
 }
 
-unsigned long Building::numCuboids()
+int Building::numCuboids()
 {
-    return this->boxes->size();
+    return (int) this->boxes->size();
 }
